Move ActionSpaces to new directory
