+++ domain/actionSpace/gatherWood.unit.ts
@@ -1,10 +1,10 @@
 import { expect } from "chai";
 import { GatherWood } from "./gatherWood";
-import { Dwarf, Player } from "./entity/Player";
-import { ActionSpace, ActionSpaceId } from "./entity/ActionSpace";
-import { Resources, ResourceType } from "./entity/Resources";
-import { Game } from "./entity/Game";
-import { buildBaseObjects, expectMutationsOfType, shouldPlaceDwarf } from "./testUtils";
+import { Dwarf, Player } from "../entity/Player";
+import { ActionSpace, ActionSpaceId } from "../entity/ActionSpace";
+import { Resources, ResourceType } from "../entity/Resources";
+import { Game } from "../entity/Game";
+import { buildBaseObjects, expectMutationsOfType, shouldPlaceDwarf } from "../testUtils";
 
 describe("Gather wood", () => {
     shouldPlaceDwarf(GatherWood.execute);
