+++ domain/actionSpace/urgentWishForChildren.unit.ts
@@ -1,8 +1,8 @@
-import { buildBaseObjects, expectMutationsOfType, shouldPlaceDwarf } from "./testUtils";
+import { buildBaseObjects, expectMutationsOfType, shouldPlaceDwarf } from "../testUtils";
 import { UrgentWishForChildren } from "./urgentWishForChildren";
-import { Dwarf, Player } from "./entity/Player";
-import { ActionSpace } from "./entity/ActionSpace";
-import { EntityMutation, EntityType, isMutationOfType } from "./entity/Mutation";
+import { Dwarf, Player } from "../entity/Player";
+import { ActionSpace } from "../entity/ActionSpace";
+import { EntityMutation, EntityType, isMutationOfType } from "../entity/Mutation";
 import { expect } from "chai";
 
 describe("Urgent Wish for Children", () => {
